+++ src/node/core/index.ts
@@ -1,6 +1,6 @@
 import path from "path";
 import chokidar from "chokidar";
-import { ResolvedUserConfig, SidebarNode } from "../types.js";
+import { Command, ResolvedUserConfig, SidebarNode } from "../types.js";
 import {
   MD_PATTERN,
   VIRTUAL_APP_DATA_ID,
@@ -15,10 +15,11 @@ import { Markdown } from "./Markdown.js";
 import { withVirtual, VirtualUpdater } from "../plugins/virtual/index.js";
 import { pkg } from "../constants.js";
 import { PluginOption } from "vite";
+import { logger } from "../utils/index.js";
 
 class Core {
   private static instance: Core;
-
+  private command: Command;
   private config: ResolvedUserConfig;
   private watcher: chokidar.FSWatcher;
   private markdownCache: MarkdownCache;
@@ -32,22 +33,9 @@ class Core {
     updateSandBoxes: VirtualUpdater;
   };
 
-  private constructor(config: ResolvedUserConfig) {
+  private constructor(command: Command, config: ResolvedUserConfig) {
     this.config = config;
-    this.watcher = chokidar.watch(path.join(config.docs, MD_PATTERN), {
-      atomic: false,
-      awaitWriteFinish: false,
-      binaryInterval: 300,
-      disableGlobbing: false,
-      followSymlinks: true,
-      ignored: ["**/node_modules/**", "**/.git/**"],
-      ignoreInitial: true,
-      ignorePermissionErrors: true,
-      interval: 100,
-      persistent: true,
-      useFsEvents: true,
-    });
-
+    this.command = command;
     this.markdownCache = new MarkdownCache(config);
     this.updater = {
       updateSidebars: () => null,
@@ -61,6 +49,21 @@ class Core {
   }
 
   private watch() {
+    const watchedPath = path.join(this.config.docs, MD_PATTERN);
+    this.watcher = chokidar.watch(watchedPath, {
+      atomic: false,
+      awaitWriteFinish: false,
+      binaryInterval: 300,
+      disableGlobbing: false,
+      followSymlinks: true,
+      ignored: ["**/node_modules/**", "**/.git/**"],
+      ignoreInitial: true,
+      ignorePermissionErrors: true,
+      interval: 100,
+      persistent: true,
+      useFsEvents: true,
+    });
+
     this.watcher.addListener("all", async (eventType: string, file: string) => {
       switch (eventType) {
         case "add": {
@@ -83,6 +86,11 @@ class Core {
         }
       }
     });
+
+    logger.info(`
+Start watching docs at 
+${watchedPath}
+`);
   }
 
   private async makeViteVirtualPlugin() {
@@ -272,10 +280,6 @@ class Core {
     `;
   }
 
-  private setConfig(config: ResolvedUserConfig) {
-    this.config = config;
-  }
-
   getTmp() {
     return this.tmp;
   }
@@ -293,19 +297,18 @@ class Core {
     const { sidebars, routes, appData, sandboxes } =
       await this.makeViteVirtualPlugin();
 
-    this.watch();
+    if (this.command === "start") {
+      this.watch();
+    }
 
     return [sidebars, routes, appData, sandboxes];
   }
 
-  static getInstance(config?: ResolvedUserConfig) {
+  static getInstance(command?: Command, config?: ResolvedUserConfig) {
     if (!Core.instance) {
-      Core.instance = new Core(config);
-    } else {
-      if (config) {
-        Core.instance.setConfig(config);
-      }
+      Core.instance = new Core(command, config);
     }
+
     return Core.instance;
   }
 }