+++ CHANGELOG.md
@@ -25,6 +25,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - optimizeDeps @mdx-js/react
 - sidebar highlight now support zhCN
 
+### Fixed
+
+- Fix chokidar still watching in build command
+
 ## [0.10.2] - 2022-08-02
 
 ### Changed